// src/components/reach-us/types.ts
export type Location = {
    _id: string;
    locationName?: string;
    verticalAssociation?: string;
    locationImage?: string;
    streetAddress?: string;
    city?: string;
    stateProvince?: string;
    zipPostalCode?: string;
    phoneNumber?: string;
    emailAddress?: string;
    mapEmbedUrl?: string;
};